fix(plugin-crx): lazy-load clip listener activate to keep CrxPlugin chunk small
The install-crx-bridge module was defined inline in CrxPlugin.tsx, pulling
listener.ts, mapping.ts, and their transitive deps (@dxos/plugin-client,
@dxos/plugin-space, @dxos/plugin-markdown, @dxos/types) into the eagerly
loaded CrxPlugin chunk. That expanded the chunk graph rooted at the plugin
and triggered an ESM init ordering issue in Linux webkit where the AppSurface
namespace import (`u`) was undefined when ReactSurface activation ran the
filter, surfacing a System Error dialog and breaking the entire app boot in
the e2e suite.

Move the bridge activate into capabilities/install-clip-listener.ts behind
Capability.lazy, matching the convention used by the other capabilities. The
CrxPlugin chunk is restored to its small pre-#11292 static graph.
