+++ packages/plugins/plugin-crx/src/CrxPlugin.tsx
@@ -2,31 +2,12 @@
 // Copyright 2026 DXOS.org
 //
 
-import * as Effect from 'effect/Effect';
+import { ActivationEvents, Plugin } from '@dxos/app-framework';
+import { AppPlugin } from '@dxos/app-toolkit';
 
-import { ActivationEvents, Capabilities, Capability, Plugin } from '@dxos/app-framework';
-import { AppPlugin, LayoutOperation } from '@dxos/app-toolkit';
-import { log } from '@dxos/log';
-
-import { CrxSettings, ReactSurface } from '#capabilities';
+import { CrxSettings, InstallClipListener, ReactSurface } from '#capabilities';
 import { meta } from '#meta';
 import { translations } from '#translations';
-
-import { installClipListener } from './listener';
-
-const SUCCESS_TOAST_BY_KIND: Record<string, string> = {
-  person: 'toast.person.title',
-  organization: 'toast.organization.title',
-  note: 'toast.note.title',
-};
-
-const ERROR_TOAST_BY_CODE: Record<string, string> = {
-  invalidPayload: 'toast.error.invalidPayload.message',
-  unsupportedVersion: 'toast.error.unsupportedVersion.message',
-  unsupportedKind: 'toast.error.unsupportedKind.message',
-  noSpace: 'toast.error.noSpace.message',
-  internal: 'toast.error.internal.message',
-};
 
 export const CrxPlugin = Plugin.define(meta).pipe(
   AppPlugin.addSettingsModule({ activate: CrxSettings }),
@@ -35,30 +16,7 @@ export const CrxPlugin = Plugin.define(meta).pipe(
   Plugin.addModule({
     id: 'install-crx-bridge',
     activatesOn: ActivationEvents.OperationInvokerReady,
-    activate: Effect.fnUntraced(function* () {
-      const capabilityManager = yield* Capability.Service;
-      const invoker = yield* Capability.get(Capabilities.OperationInvoker);
-
-      installClipListener(capabilityManager, invoker, (ack, detail) => {
-        const kind = (detail as { kind?: string } | null)?.kind;
-        if (ack.ok) {
-          const key = (kind && SUCCESS_TOAST_BY_KIND[kind]) ?? 'toast.person.title';
-          void invoker.invokePromise(LayoutOperation.AddToast, {
-            id: `${meta.id}.clip-${ack.id}`,
-            title: [key, { ns: meta.id }],
-          });
-        } else {
-          const errorKey = ERROR_TOAST_BY_CODE[ack.error] ?? 'toast.error.internal.message';
-          void invoker.invokePromise(LayoutOperation.AddToast, {
-            id: `${meta.id}.error-${Date.now()}`,
-            title: ['toast.error.title', { ns: meta.id }],
-            description: [errorKey, { ns: meta.id }],
-          });
-        }
-      });
-
-      log.info('CRX bridge installed');
-    }),
+    activate: InstallClipListener,
   }),
   Plugin.make,
 );
