+++ packages/plugins/plugin-crx/src/capabilities/index.ts
@@ -5,4 +5,5 @@
 import { Capability } from '@dxos/app-framework';
 
 export const CrxSettings = Capability.lazy('CrxSettings', () => import('./settings'));
+export const InstallClipListener = Capability.lazy('InstallClipListener', () => import('./install-clip-listener'));
 export const ReactSurface = Capability.lazy('ReactSurface', () => import('./react-surface'));
